+++ MonsterClass.py
@@ -4,7 +4,7 @@ import random
 class Monster:
     # The monster class defines a name, position and whether the monster is currently hidden on the board
 
-    def __init__(self, name, symbol, position, hidden, found, hp, attack, defence, defeated, gold_loot):
+    def __init__(self, name, symbol, position, hidden, found, hp, attack, defence, defeated, gold_loot, xp):
         self.name = name
         self.symbol = symbol
         self.position = position
@@ -15,6 +15,7 @@ class Monster:
         self.defence = defence
         self.defeated = defeated
         self.gold_loot = gold_loot
+        self.xp = xp
 
 
 # Generate random amount of gold for orcs
@@ -30,6 +31,6 @@ def gen_ran_pos():
 
 
 # Create 10 orcs on the board
-army_of_orcs = [Monster("Bald Orc", "m", gen_ran_pos(), " ", False, 100, 11, 2, False, gen_orc_gold()) for _ in range(10)]
+army_of_orcs = [Monster("Bald Orc", "m", gen_ran_pos(), " ", False, 100, 32, 2, False, gen_orc_gold(), 49) for _ in range(10)]
 
-orc_boss = Monster("Destroyer Orc", "B", 99, "B", False, 500, 30, 10, False, 300)
+orc_boss = Monster("Destroyer Orc", "B", 99, "B", False, 500, 65, 10, False, 300, 100)